+++ TestUtils/src/main/java/com/braintreepayments/api/Fixtures.kt
@@ -1,5 +1,6 @@
 package com.braintreepayments.api
 
+@Suppress("LargeClass")
 object Fixtures {
 
     // region 3D Secure
@@ -408,9 +409,21 @@ object Fixtures {
     // endregion
 
     // region Auth Tokens
-    const val BASE64_CLIENT_TOKEN = "eyJ2ZXJzaW9uIjoyLCJhdXRob3JpemF0aW9uRmluZ2VycHJpbnQiOiJlbmNvZGVkX2F1dGhfZmluZ2VycHJpbnQiLCJjaGFsbGVuZ2VzIjpbImN2diIsInBvc3RhbF9jb2RlIl0sImNvbmZpZ1VybCI6ImVuY29kZWRfY2FwaV9jb25maWd1cmF0aW9uX3VybCIsImFzc2V0c1VybCI6Imh0dHA6Ly9sb2NhbGhvc3Q6OTAwMCIsInRocmVlRFNlY3VyZUVuYWJsZWQiOmZhbHNlLCJwYXlwYWxFbmFibGVkIjpmYWxzZX0="
-    const val BASE64_CLIENT_TOKEN2 = "eyJ2ZXJzaW9uIjoyLCJhdXRob3JpemF0aW9uRmluZ2VycHJpbnQiOiJlbmNvZGVkX2F1dGhfZmluZ2VycHJpbnQiLCJjaGFsbGVuZ2VzIjpbImN2diIsInBvc3RhbF9jb2RlIl0sImNvbmZpZ1VybCI6ImVuY29kZWRfY2FwaV9jb25maWd1cmF0aW9uX3VybCIsImFzc2V0c1VybCI6Imh0dHA6Ly9sb2NhbGhvc3Q6OTAwMiIsInRocmVlRFNlY3VyZUVuYWJsZWQiOmZhbHNlLCJwYXlwYWxFbmFibGVkIjpmYWxzZX0="
-    const val BASE64_CLIENT_TOKEN_WITH_SPACES = " eyJ2ZXJzaW9uIjoyLCJhdXRob3JpemF0aW9uRmluZ2VycHJpbnQiOiJlbmNvZGVkX2F1dGhfZmluZ2VycHJpbnQiLCJjaGFsbGVuZ2VzIjpbImN2diIsInBvc3RhbF9jb2RlIl0sImNvbmZpZ1VybCI6ImVuY29kZWRfY2FwaV9jb25maWd1cmF0aW9uX3VybCIsImFzc2V0c1VybCI6Imh0dHA6Ly9sb2NhbGhvc3Q6OTAwMCIsInRocmVlRFNlY3VyZUVuYWJsZWQiOmZhbHNlLCJwYXlwYWxFbmFibGVkIjpmYWxzZX0= "
+    const val BASE64_CLIENT_TOKEN = "eyJ2ZXJzaW9uIjoyLCJhdXRob3JpemF0aW9uRmluZ2VycHJpbnQiOiJl" +
+            "bmNvZGVkX2F1dGhfZmluZ2VycHJpbnQiLCJjaGFsbGVuZ2VzIjpbImN2diIsInBvc3RhbF9jb2RlIl0sIm" +
+            "NvbmZpZ1VybCI6ImVuY29kZWRfY2FwaV9jb25maWd1cmF0aW9uX3VybCIsImFzc2V0c1VybCI6Imh0dHA6" +
+            "Ly9sb2NhbGhvc3Q6OTAwMCIsInRocmVlRFNlY3VyZUVuYWJsZWQiOmZhbHNlLCJwYXlwYWxFbmFibGVkIjp" +
+            "mYWxzZX0="
+    const val BASE64_CLIENT_TOKEN2 = "eyJ2ZXJzaW9uIjoyLCJhdXRob3JpemF0aW9uRmluZ2VycHJpbnQiOiJlb" +
+            "mNvZGVkX2F1dGhfZmluZ2VycHJpbnQiLCJjaGFsbGVuZ2VzIjpbImN2diIsInBvc3RhbF9jb2RlIl0sImN" +
+            "vbmZpZ1VybCI6ImVuY29kZWRfY2FwaV9jb25maWd1cmF0aW9uX3VybCIsImFzc2V0c1VybCI6Imh0dHA6L" +
+            "y9sb2NhbGhvc3Q6OTAwMiIsInRocmVlRFNlY3VyZUVuYWJsZWQiOmZhbHNlLCJwYXlwYWxFbmFibGVkIjp" +
+            "mYWxzZX0="
+    const val BASE64_CLIENT_TOKEN_WITH_SPACES = " eyJ2ZXJzaW9uIjoyLCJhdXRob3JpemF0aW9uRmluZ2Vyc" +
+            "HJpbnQiOiJlbmNvZGVkX2F1dGhfZmluZ2VycHJpbnQiLCJjaGFsbGVuZ2VzIjpbImN2diIsInBvc3RhbF9" +
+            "jb2RlIl0sImNvbmZpZ1VybCI6ImVuY29kZWRfY2FwaV9jb25maWd1cmF0aW9uX3VybCIsImFzc2V0c1Vyb" +
+            "CI6Imh0dHA6Ly9sb2NhbGhvc3Q6OTAwMCIsInRocmVlRFNlY3VyZUVuYWJsZWQiOmZhbHNlLCJwYXlwYWx" +
+            "FbmFibGVkIjpmYWxzZX0= "
 
     const val TOKENIZATION_KEY = "sandbox_tmxhyf7d_dcpspy2brwdjr3qn"
     const val TOKENIZATION_KEY_WITH_SPACES = " sandbox_tmxhyf7d_dcpspy2brwdjr3qn "
@@ -2435,6 +2448,83 @@ object Fixtures {
         } 
     """
 
+    // language=JSON
+    const val PAYMENT_METHODS_GOOGLE_PAY_REQUEST_NO_CREDIT_CARDS = """
+        {
+          "apiVersion": 2,
+          "apiVersionMinor": 0,
+          "allowedPaymentMethods": [
+            {
+              "type": "CARD",
+              "parameters": {
+                "billingAddressRequired": true, 
+                "allowPrepaidCards": true,
+                "allowCreditCards": true,
+                "allowedAuthMethods": [
+                  "PAN_ONLY",
+                  "CRYPTOGRAM_3DS"
+                ],
+                "allowedCardNetworks": [
+                  "VISA",
+                  "AMEX",
+                  "JCB",
+                  "DISCOVER",
+                  "MASTERCARD"
+                ],
+                "allowCreditCards": false
+              },
+              "tokenizationSpecification": {
+                "type": "PAYMENT_GATEWAY",
+                "parameters": {
+                  "gateway": "braintree",
+                  "braintree:apiVersion": "v1",
+                  "braintree:sdkVersion": "BETA",
+                  "braintree:merchantId": "BRAINTREE_MERCHANT_ID",
+                  "braintree:authorizationFingerprint": "BRAINTREE_AUTH_FINGERPRINT"
+                }
+              }
+            },
+            {
+              "type": "PAYPAL",
+              "parameters": {
+                "purchase_context": "{\"purchase_context\":{\"purchase_units\":[{\"payee\":{\"client_id\":\"FAKE_PAYPAL_CLIENT_ID\"},\"recurring_payment\":false}]}}"
+              },
+              "tokenizationSpecification": {
+                "type": "PAYMENT_GATEWAY",
+                "parameters": {
+                  "gateway": "braintree",
+                  "braintree:apiVersion": "v1",
+                  "braintree:sdkVersion": "BETA",
+                  "braintree:merchantId": "BRAINTREE_MERCHANT_ID",
+                  "braintree:authorizationFingerprint": "BRAINTREE_AUTH_FINGERPRINT"
+                }
+              }
+            }
+          ],
+          "emailRequired": true,
+          "shippingAddressRequired": true,
+          "shippingAddressParameters": {
+            "allowedCountryCodes": [
+              "US",
+              "CA",
+              "MX",
+              "GB"
+            ],
+            "phoneNumberRequired": true
+          },
+          "environment": "PRODUCTION",
+          "merchantInfo": {
+            "merchantId": "GOOGLE_MERCHANT_ID",
+            "merchantName": "GOOGLE_MERCHANT_NAME"
+          },
+          "transactionInfo": {
+            "totalPriceStatus": "FINAL",
+            "totalPrice": "12.24",
+            "currencyCode": "USD",
+            "countryCode": "US"
+          }
+        } 
+    """
 
     // language=JSON
     const val RESPONSE_GOOGLE_PAY_CARD = """